requirements.txt
pyyaml==6.0.2
requests==2.32.3
torch==2.4.0
torchvision==0.19.0
tensorboard==2.17.1
pycocotools==2.0.8
notebook==7.2.2
tqdm==4.66.5
